Modify Projucer Android exporter and java files for JuceBridge class
