+++ modules/juce_core/native/java/AndroidRuntimePermissions.java
@@ -1,4 +1,3 @@
-    @Override
     public void onRequestPermissionsResult (int permissionID, String permissions[], int[] grantResults)
     {
         boolean permissionsGranted = (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED);
