+++ modules/juce_core/native/juce_android_JNIHelpers.h
@@ -29,10 +29,6 @@
 #ifndef JUCE_ANDROID_JNIHELPERS_H_INCLUDED
 #define JUCE_ANDROID_JNIHELPERS_H_INCLUDED
 
-#if ! (defined (JUCE_ANDROID_ACTIVITY_CLASSNAME) && defined (JUCE_ANDROID_ACTIVITY_CLASSPATH))
- #error "The JUCE_ANDROID_ACTIVITY_CLASSNAME and JUCE_ANDROID_ACTIVITY_CLASSPATH macros must be set!"
-#endif
-
 #if ! (defined (JUCE_ANDROID_BRIDGE_CLASSNAME) && defined (JUCE_ANDROID_BRIDGE_CLASSPATH))
  #error "The JUCE_ANDROID_BRIDGE_CLASSNAME and JUCE_ANDROID_BRIDGE_CLASSPATH macros must be set!"
 #endif
