+++ modules/juce_gui_basics/native/juce_android_Windowing.cpp
@@ -135,15 +135,14 @@ public:
           sizeAllocated (0),
           scale ((float) Desktop::getInstance().getDisplays().getMainDisplay().scale)
     {
-        DBG ("AndroidComponentPeer constructor");
         // NB: must not put this in the initialiser list, as it invokes a callback,
         // which will fail if the peer is only half-constructed.
         DBG ("AndroidComponentPeer constructor: "+component.getName());
-        jstring componentName = javaString(component.getName());
         view = GlobalRef (android.bridge.callObjectMethod (JuceBridge.createNewView,
                                                            (jboolean) component.isOpaque(),
                                                            (jlong) this,
                 (jstring) javaString(component.getName())));
+
 
         if (isFocused())
             handleFocusGain();
